feat: Rewrote the CCC web client
