+++ ccc-client/src/plugins/vuetify.ts
@@ -1,0 +1,33 @@
+/**
+ * plugins/vuetify.ts
+ *
+ * Framework documentation: https://vuetifyjs.com`
+ */
+
+// Styles
+import '@mdi/font/css/materialdesignicons.css'
+import 'vuetify/styles'
+
+// Composables
+import { createVuetify } from 'vuetify'
+import { VStepper, VStepperHeader, VStepperItem } from 'vuetify/labs/VStepper'
+
+// https://vuetifyjs.com/en/introduction/why-vuetify/#feature-guides
+export default createVuetify({
+  components: {
+    VStepper,
+    VStepperHeader,
+    VStepperItem
+  },
+  theme: {
+    themes: {
+      light: {
+        colors: {
+          primary: '#1867C0',
+          secondary: '#5CBBF6',
+
+        },
+      },
+    },
+  },
+})
