--- ccc/client/package.json
@@ -1,63 +1,0 @@
-{
-  "name": "ccc",
-  "version": "3.1.5",
-  "private": true,
-  "engines": {
-    "node": ">=14 <=18",
-    "npm": ">=8.3.0 <=9"
-  },
-  "scripts": {
-    "serve": "vue-cli-service serve",
-    "build": "vue-cli-service build",
-    "test:unit": "vue-cli-service test:unit",
-    "test:e2e": "vue-cli-service test:e2e",
-    "lint": "vue-cli-service lint"
-  },
-  "dependencies": {
-    "axios": "^0.27.2",
-    "core-js": "^3.6.5",
-    "vue": "^2.7.10",
-    "vue-class-component": "^7.2.3",
-    "vue-markdown-render": "^1.1.3",
-    "vue-property-decorator": "^8.4.2",
-    "vuetify": "^2.2.11"
-  },
-  "devDependencies": {
-    "@types/chai": "^4.2.11",
-    "@types/mocha": "^5.2.7",
-    "@vue/cli-plugin-babel": "~5.0.8",
-    "@vue/cli-plugin-e2e-cypress": "~5.0.8",
-    "@vue/cli-plugin-typescript": "~5.0.8",
-    "@vue/cli-plugin-unit-mocha": "^5.0.8",
-    "@vue/cli-service": "~5.0.8",
-    "@vue/test-utils": "^1.3.0",
-    "chai": "^4.3.6",
-    "cypress": "^10.7.0",
-    "lorem-ipsum": "^2.0.8",
-    "miragejs": "^0.1.45",
-    "sass": "^1.54.9",
-    "sass-loader": "^13.0.2",
-    "typescript": "~4.8.3",
-    "vue-cli-plugin-vuetify": "~2.5.5",
-    "vue-template-compiler": "^2.7.10",
-    "vuetify-loader": "^1.3.0"
-  },
-  "overrides": {
-    "ansi-regex": "^5.0.1",
-    "follow-redirects": "^1.15.2",
-    "glob-parent": "^6.0.2",
-    "hosted-git-info": "^2.8.9",
-    "ini": "^1.3.6",
-    "jsdom": "^18.0.1",
-    "lodash": "^4.17.21",
-    "markdown-it": "^13.0.1",
-    "minimist": ">=1.2.6",
-    "nanoid": "^3.3.4",
-    "node-forge": "^1.3.1",
-    "postcss": ">=8.4.16",
-    "shelljs": "^0.8.5",
-    "url-parse": "^1.5.10",
-    "ws": "^8.8.1",
-    "y18n": "^4.0.1"
-  }
-}
